🩹 fix: - 주소에 /api 들 제거 - board 클릭 시 board 페이지로 이동하게 추가 - board에서 boardID 표기 방법 추가 - 프로필 수정 버튼 추가 및 페이지 연결
