+++ src/main/resources/templates/index.html
@@ -14,7 +14,7 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/twbs-pagination/1.4.2/jquery.twbsPagination.min.js"></script>
 
 
-    <script src="/js/basic.js"></script>
+    <script src="/js/basic_Index.js"></script>
     <title>메인 페이지</title>
     <style>
         .header {
@@ -63,6 +63,8 @@
     <div id="header-title-select-shop">
         Kanban Board
     </div>
+    <button type="button" class="btn btn-secondary mb-2" id="login-id-btn" onclick="location.href='/members/myInfo'">유저 프로필</button>
+    <br/>
     <a id="login-text" href="javascript:logout()">
         로그아웃
     </a>
