+++ src/main/resources/templates/myInfo.html
@@ -10,6 +10,8 @@
   <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
   <!-- js-cookie -->
   <script src="https://cdn.jsdelivr.net/npm/js-cookie@3.0.5/dist/js.cookie.min.js"></script>
+
+  <script src="/js/basic.js"></script>
   <style>
     body {
       font-family: Arial, sans-serif;
@@ -177,49 +179,6 @@
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"></script>
 </body>
 <script>
-  $(document).ready(function () {
-    // 토큰 삭제
-    Cookies.remove('Authorization', {path: '/'});
-  });
-
-  var myModal = new bootstrap.Modal(document.getElementById('exampleModal'));
-
-  const href = location.href;
-  const queryString = href.substring(href.indexOf("?") + 1)
-  if (queryString === 'error') {
-    myModal.show();
-  }
-
-  const host = 'http://' + window.location.host;
-
-  function onLogin() {
-    let username = $('#username').val();
-    let password = $('#password').val();
-
-    $.ajax({
-      type: "POST",
-      url: `/api/members/login`,
-      contentType: "application/json",
-      data: JSON.stringify({email: username, password: password}),
-    })
-    .done(function (res, status, xhr) {
-      alert("로그인 성공");
-      const token = xhr.getResponseHeader('Authorization');
-
-      Cookies.set('Authorization', token, {path: '/'})
-
-      $.ajaxPrefilter(function (options, originalOptions, jqXHR) {
-        jqXHR.setRequestHeader('Authorization', token);
-      });
-
-      window.location.href = host;
-    })
-    .fail(function (jqXHR, textStatus) {
-      myModal.show();
-      window.location.href = host + '/api/members/login-page?error'
-    });
-  }
-
   function findPassword() {
     alert("개발 중입니다");
   }
