+++ src/main/resources/templates/test.html
@@ -1,0 +1,63 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Title</title>
+    <!-- Bootstrap CSS -->
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet">
+    <!-- jQuery -->
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
+    <!-- js-cookie -->
+    <script src="https://cdn.jsdelivr.net/npm/js-cookie@3.0.5/dist/js.cookie.min.js"></script>
+</head>
+<body>
+<h1>Board ID: <span th:text="${boardId}"></span></h1>
+</body>
+
+<script>
+    $(document).ready(function () {
+        const auth = getToken();
+
+        if (auth !== undefined && auth !== '') {
+            $.ajaxPrefilter(function (options, originalOptions, jqXHR) {
+                jqXHR.setRequestHeader('Authorization', auth);
+            });
+        } else {
+            window.location.href = host + '/members/login-page';
+            return;
+        }
+
+        $.ajax({
+            type: 'GET',
+            url: `/api/members`,
+            contentType: 'application/json',
+        })
+            .done(function (res, status, xhr) {
+                const username = res.data.email;
+                const isAdmin = !!res.admin;
+
+                if (!username) {
+                    alert("여기");
+                    window.location.href = '/api/members/login-page';
+                    return;
+                }
+
+                $('#username').text(username);
+                showBoard();
+                /*
+                if (isAdmin) {
+                    $('#admin').text(true);
+                    showProduct();
+                } else {
+                    showProduct();
+                }
+                */
+
+            })
+            .fail(function (jqXHR, textStatus) {
+                logout();
+            });
+
+    })
+</script>
+</html>
